Separate archive api into separate package
