+++ radixdlt-core/radixdlt/src/integration/java/com/radixdlt/integration/staking/StakingUnstakingValidatorsTest.java
@@ -67,10 +67,10 @@ package com.radixdlt.integration.staking;
 import com.google.common.collect.ClassToInstanceMap;
 import com.google.inject.Scopes;
 import com.google.inject.multibindings.Multibinder;
-import com.radixdlt.api.accounts.BerkeleyAccountInfoStore;
+import com.radixdlt.api.archive.accounts.BerkeleyAccountInfoStore;
 import com.radixdlt.api.store.ValidatorUptime;
 import com.radixdlt.api.store.berkeley.BerkeleyValidatorUptimeArchiveStore;
-import com.radixdlt.api.tokens.BerkeleyResourceInfoStore;
+import com.radixdlt.api.archive.tokens.BerkeleyResourceInfoStore;
 import com.radixdlt.application.validators.scrypt.ValidatorUpdateRakeConstraintScrypt;
 import com.radixdlt.consensus.LedgerProof;
 import com.radixdlt.consensus.bft.View;
