+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/archive/ArchiveEndpointModule.java
@@ -1,10 +1,9 @@
-/* Copyright 2021 Radix Publishing Ltd incorporated in Jersey (Channel Islands).
- *
+/*
+ * Copyright 2021 Radix DLT Ltd incorporated in England.
  * Licensed under the Radix License, Version 1.0 (the "License"); you may not use this
  * file except in compliance with the License. You may obtain a copy of the License at:
  *
  * radixfoundation.org/licenses/LICENSE-v1
- *
  * The Licensor hereby grants permission for the Canonical version of the Work to be
  * published, distributed and used under or by reference to the Licensor’s trademark
  * Radix ® and use of any unregistered trade names, logos or get-up.
@@ -62,7 +61,7 @@
  * permissions under this License.
  */
 
-package com.radixdlt.api.module;
+package com.radixdlt.api.archive;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
@@ -70,15 +69,15 @@ import com.google.inject.multibindings.ProvidesIntoMap;
 import com.google.inject.multibindings.StringMapKey;
 import com.radixdlt.api.Controller;
 import com.radixdlt.api.JsonRpcHandler;
-import com.radixdlt.api.accounts.AccountApiModule;
+import com.radixdlt.api.archive.accounts.AccountApiModule;
+import com.radixdlt.api.archive.transactions.TransactionStatusAndLookupApiModule;
 import com.radixdlt.api.controller.JsonRpcController;
 import com.radixdlt.api.handler.ArchiveNetworkHandler;
-import com.radixdlt.api.transactions.TransactionApiModule;
 import com.radixdlt.api.qualifier.ArchiveEndpoint;
 import com.radixdlt.api.qualifier.ArchiveServer;
 import com.radixdlt.api.server.JsonRpcServer;
-import com.radixdlt.api.tokens.TokenApiModule;
-import com.radixdlt.api.validators.ValidatorApiModule;
+import com.radixdlt.api.archive.tokens.TokenApiModule;
+import com.radixdlt.api.archive.validators.ValidatorApiModule;
 
 import java.util.Map;
 
@@ -87,7 +86,7 @@ public class ArchiveEndpointModule extends AbstractModule {
 	protected void configure() {
 		install(new AccountApiModule());
 		install(new TokenApiModule());
-		install(new TransactionApiModule());
+		install(new TransactionStatusAndLookupApiModule());
 		install(new ValidatorApiModule());
 	}
