+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/archive/accounts/BerkeleyAccountInfoStore.java
@@ -61,7 +61,7 @@
  * permissions under this License.
  */
 
-package com.radixdlt.api.accounts;
+package com.radixdlt.api.archive.accounts;
 
 import com.google.inject.Inject;
 import com.google.inject.Provider;
