+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/transactions/TransactionIndexApiModule.java
@@ -61,7 +61,7 @@
  * permissions under this License.
  */
 
-package com.radixdlt.api.transactions.index;
+package com.radixdlt.api.transactions;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Scopes;
@@ -70,7 +70,7 @@ import com.google.inject.multibindings.ProvidesIntoMap;
 import com.google.inject.multibindings.StringMapKey;
 import com.radixdlt.api.JsonRpcHandler;
 import com.radixdlt.api.qualifier.ArchiveEndpoint;
-import com.radixdlt.api.transactions.lookup.BerkeleyTransactionsByIdStore;
+import com.radixdlt.api.qualifier.TransactionsEndpoint;
 import com.radixdlt.store.berkeley.BerkeleyAdditionalStore;
 import com.radixdlt.utils.functional.Failure;
 import com.radixdlt.utils.functional.Result;
@@ -89,9 +89,9 @@ public final class TransactionIndexApiModule extends AbstractModule {
 			.addBinding().to(BerkeleyTransactionIndexStore.class);
 	}
 
-	@ArchiveEndpoint
+	@TransactionsEndpoint
 	@ProvidesIntoMap
-	@StringMapKey("transactions.get_count")
+	@StringMapKey("get_transaction_count")
 	public JsonRpcHandler indexGetTransactionCount(BerkeleyTransactionIndexStore store) {
 		return request -> withRequiredParameters(
 			request,
@@ -107,9 +107,9 @@ public final class TransactionIndexApiModule extends AbstractModule {
 		);
 	}
 
-	@ArchiveEndpoint
+	@TransactionsEndpoint
 	@ProvidesIntoMap
-	@StringMapKey("transactions.get_transactions")
+	@StringMapKey("get_transactions")
 	public JsonRpcHandler indexGetTransactions(BerkeleyTransactionIndexStore store, BerkeleyTransactionsByIdStore txnStore) {
 		return request -> withRequiredParameters(
 			request,
